requirements.txt
anyio==3.6.2
autoflake==2.0.0
bandit==1.7.4
black==22.12.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
et-xmlfile==1.1.0
fastapi==0.88.0
filelock==3.8.2
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.29
h11==0.14.0
identify==2.5.9
idna==3.4
isort==5.11.2
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
pathspec==0.10.3
pbr==5.11.0
platformdirs==2.6.0
pre-commit==2.20.0
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.22.0
stevedore==4.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.17.1
